fix: Move derived to custom collector
