+++ src/capellambse_context_diagrams/collectors/custom.py
@@ -5,13 +5,32 @@
 
 from __future__ import annotations
 
+import collections.abc as cabc
 import copy
 import typing as t
 
 import capellambse.model as m
+from capellambse.metamodel import cs, fa, la, sa
 
 from .. import _elkjs, context
-from . import generic, makers
+from . import default, generic, makers
+
+if t.TYPE_CHECKING:
+    from .. import context
+
+    DerivatorFunction: t.TypeAlias = cabc.Callable[
+        [
+            context.CustomDiagram,
+            _elkjs.ELKInputData,
+            dict[str, _elkjs.ELKInputChild],
+        ],
+        None,
+    ]
+
+    Filter: t.TypeAlias = cabc.Callable[
+        [cabc.Iterable[m.ModelElement]],
+        cabc.Iterable[m.ModelElement],
+    ]
 
 
 def _is_edge(obj: m.ModelElement) -> bool:
@@ -93,6 +112,14 @@ class CustomCollector:
             for edge_uuid, box_uuid in self.edge_owners.items():
                 if box := self.boxes.get(box_uuid):
                     box.edges.append(self.edges.pop(edge_uuid))
+
+        derivator = DERIVATORS.get(type(self.target))
+        if self.diagram._display_derived_interfaces and derivator is not None:
+            derivator(
+                self.diagram,
+                self.data,
+                self.boxes,
+            )
 
         self._fix_box_heights()
         for uuid in self.boxes_to_delete:
@@ -292,3 +319,84 @@ def collector(
 ) -> _elkjs.ELKInputData:
     """Collect data for rendering a custom diagram."""
     return CustomCollector(diagram, params)()
+
+
+def derive_from_functions(
+    diagram: context.CustomDiagram,
+    data: _elkjs.ELKInputData,
+    boxes: dict[str, _elkjs.ELKInputChild],
+) -> None:
+    """Derive Components from allocated functions of the context target.
+
+    A Component, a ComponentExchange and two ComponentPorts are added
+    to ``data``. These elements are prefixed with ``Derived-`` to
+    receive special styling in the serialization step.
+    """
+    assert isinstance(diagram.target, cs.Component)
+    ports: list[m.ModelElement] = []
+    for fnc in diagram.target.allocated_functions:
+        inc, out = default.port_collector(fnc, diagram.type)
+        ports.extend((inc | out).values())
+
+    derived_components: dict[str, cs.Component] = {}
+    for port in ports:
+        for fex in port.exchanges:
+            if isinstance(port, fa.FunctionOutputPort):
+                attr = "target"
+            else:
+                attr = "source"
+
+            try:
+                derived_comp = getattr(fex, attr).owner.owner
+                if (
+                    derived_comp == diagram.target
+                    or derived_comp.uuid in boxes
+                ):
+                    continue
+
+                if derived_comp.uuid not in derived_components:
+                    derived_components[derived_comp.uuid] = derived_comp
+            except AttributeError:  # No owner of owner.
+                pass
+
+    # Idea: Include flow direction of derived interfaces from all functional
+    # exchanges. Mixed means bidirectional. Just even out bidirectional
+    # interfaces and keep flow direction of others.
+
+    centerbox = boxes[diagram.target.uuid]
+    i = 0
+    for i, (uuid, derived_component) in enumerate(
+        derived_components.items(), 1
+    ):
+        box = makers.make_box(
+            derived_component,
+            no_symbol=diagram._display_symbols_as_boxes,
+        )
+        class_ = diagram.serializer.get_styleclass(derived_component.uuid)
+        box.id = f"{makers.STYLECLASS_PREFIX}-{class_}:{uuid}"
+        data.children.append(box)
+        source_id = f"{makers.STYLECLASS_PREFIX}-CP_INOUT:{i}"
+        target_id = f"{makers.STYLECLASS_PREFIX}-CP_INOUT:{-i}"
+        box.ports.append(makers.make_port(source_id))
+        centerbox.ports.append(makers.make_port(target_id))
+        if i % 2 == 0:
+            source_id, target_id = target_id, source_id
+
+        data.edges.append(
+            _elkjs.ELKInputEdge(
+                id=f"{makers.STYLECLASS_PREFIX}-ComponentExchange:{i}",
+                sources=[source_id],
+                targets=[target_id],
+            )
+        )
+
+    centerbox.height += (
+        makers.PORT_PADDING + (makers.PORT_SIZE + makers.PORT_PADDING) * i // 2
+    )
+
+
+DERIVATORS: dict[type[m.ModelElement], DerivatorFunction] = {
+    la.LogicalComponent: derive_from_functions,
+    sa.SystemComponent: derive_from_functions,
+}
+"""Supported objects to build derived contexts for."""
