+++ src/capellambse_context_diagrams/collectors/default.py
@@ -13,23 +13,14 @@ import collections.abc as cabc
 import typing as t
 
 import capellambse.model as m
-from capellambse.metamodel import cs, fa, la, sa
+from capellambse.metamodel import cs, fa
 from capellambse.model import DiagramType as DT
 
 from .. import _elkjs
-from . import custom, generic, makers
+from . import custom, generic
 
 if t.TYPE_CHECKING:
     from .. import context
-
-    DerivatorFunction: t.TypeAlias = cabc.Callable[
-        [
-            context.ContextDiagram,
-            _elkjs.ELKInputData,
-            dict[str, _elkjs.ELKInputChild],
-        ],
-        None,
-    ]
 
     Filter: t.TypeAlias = cabc.Callable[
         [cabc.Iterable[m.ModelElement]],
@@ -44,7 +35,7 @@ def collector(
     visited: set[str] = set()
     edges: set[str] = set()
 
-    def _default_collect(
+    def _collect(
         target: m.ModelElement,
     ) -> cabc.Iterator[m.ModelElement]:
         if target.uuid in visited:
@@ -64,18 +55,11 @@ def collector(
                 edges.add(exchange.uuid)
                 yield exchange
         for cmp in list(getattr(target, "components", [])):
-            yield from _default_collect(cmp)
+            yield from _collect(cmp)
 
-    diagram._collect = _default_collect(diagram.target)
+    diagram._collect = _collect(diagram.target)
     processor = custom.CustomCollector(diagram, params=params)
     processor()
-    derivator = DERIVATORS.get(type(diagram.target))
-    if diagram._display_derived_interfaces and derivator is not None:
-        derivator(
-            diagram,
-            processor.data,
-            processor.boxes,
-        )
     return processor.data
 
 
@@ -158,84 +142,3 @@ class ContextInfo(t.NamedTuple):
     This list only contains ports that at least one of the exchanges
     passed into ``collect_exchanges`` sees.
     """
-
-
-def derive_from_functions(
-    diagram: context.ContextDiagram,
-    data: _elkjs.ELKInputData,
-    boxes: dict[str, _elkjs.ELKInputChild],
-) -> None:
-    """Derive Components from allocated functions of the context target.
-
-    A Component, a ComponentExchange and two ComponentPorts are added
-    to ``data``. These elements are prefixed with ``Derived-`` to
-    receive special styling in the serialization step.
-    """
-    assert isinstance(diagram.target, cs.Component)
-    ports: list[m.ModelElement] = []
-    for fnc in diagram.target.allocated_functions:
-        inc, out = port_collector(fnc, diagram.type)
-        ports.extend((inc | out).values())
-
-    derived_components: dict[str, cs.Component] = {}
-    for port in ports:
-        for fex in port.exchanges:
-            if isinstance(port, fa.FunctionOutputPort):
-                attr = "target"
-            else:
-                attr = "source"
-
-            try:
-                derived_comp = getattr(fex, attr).owner.owner
-                if (
-                    derived_comp == diagram.target
-                    or derived_comp.uuid in boxes
-                ):
-                    continue
-
-                if derived_comp.uuid not in derived_components:
-                    derived_components[derived_comp.uuid] = derived_comp
-            except AttributeError:  # No owner of owner.
-                pass
-
-    # Idea: Include flow direction of derived interfaces from all functional
-    # exchanges. Mixed means bidirectional. Just even out bidirectional
-    # interfaces and keep flow direction of others.
-
-    centerbox = boxes[diagram.target.uuid]
-    i = 0
-    for i, (uuid, derived_component) in enumerate(
-        derived_components.items(), 1
-    ):
-        box = makers.make_box(
-            derived_component,
-            no_symbol=diagram._display_symbols_as_boxes,
-        )
-        class_ = diagram.serializer.get_styleclass(derived_component.uuid)
-        box.id = f"{makers.STYLECLASS_PREFIX}-{class_}:{uuid}"
-        data.children.append(box)
-        source_id = f"{makers.STYLECLASS_PREFIX}-CP_INOUT:{i}"
-        target_id = f"{makers.STYLECLASS_PREFIX}-CP_INOUT:{-i}"
-        box.ports.append(makers.make_port(source_id))
-        centerbox.ports.append(makers.make_port(target_id))
-        if i % 2 == 0:
-            source_id, target_id = target_id, source_id
-
-        data.edges.append(
-            _elkjs.ELKInputEdge(
-                id=f"{makers.STYLECLASS_PREFIX}-ComponentExchange:{i}",
-                sources=[source_id],
-                targets=[target_id],
-            )
-        )
-
-    centerbox.height += (
-        makers.PORT_PADDING + (makers.PORT_SIZE + makers.PORT_PADDING) * i // 2
-    )
-
-
-DERIVATORS: dict[type[m.ModelElement], DerivatorFunction] = {
-    la.LogicalComponent: derive_from_functions,
-    sa.SystemComponent: derive_from_functions,
-}
-"""Supported objects to build derived contexts for."""
